+++ cmd/acc/command/test.go
@@ -1,7 +1,7 @@
 package command
 
 import (
-	"log"
+	"errors"
 
 	"github.com/kzmshrt/acc"
 	"github.com/urfave/cli/v2"
@@ -9,14 +9,15 @@ import (
 
 func Test(c *cli.Context) error {
 	if c.NArg() < 2 {
-		log.Fatal(c.Command.UsageText)
+		return errors.New(c.Command.UsageText)
 	}
 
-	filename, url := c.Args().Get(0), c.Args().Get(1)
+	filename := c.Args().Get(0)
+	url := c.Args().Get(1)
 
-	err := acc.Test(url, filename)
+	err := acc.Test(filename, url)
 	if err != nil {
-		log.Fatal(err)
+		return err
 	}
 
 	return nil
